bot/requirements.txt
aiogram==3.22.0
aiofiles
python-dotenv
requests
beautifulsoup4
aiohttp

